Implemented Authentication and also added Redux-toolkit.

+++ src/services/CategoryService.js
@@ -4,7 +4,10 @@ import { toast } from 'react-toastify';
 export const addCategory = async (category) => {
 
     try {
-        const response = await api.post(`/categories/createcategory`, category);
+        const response = await api.post(`/categories/admin/createcategory`, category, {headers:{
+            'Authorization': `Bearer ${localStorage.getItem("accessToken")}`
+        }});
+
         toast.success('Category added successfully!');
         console.log('Category added:', response);
         return response.data;
@@ -15,9 +18,27 @@ export const addCategory = async (category) => {
     }
 }
 
+export const getAllCategoriesOfCompany = async(companyId)=>{
+    console.log("company id:" , companyId);
+    try{
+        const response = await api.get(`/categories/getallcategoriesbycompanyid/${companyId}`,{headers:{
+            'Authorization': `Bearer ${localStorage.getItem("accessToken")}`
+        }});
+
+        return response.data.data;
+    }
+    catch{
+        toast.error("Errro fecthing categories..")
+        console.log(error)
+    }
+}
+
 export const getAllCategories = async () => {
     try {
-        const response = await api.get(`/categories/getallcategories`);
+        const response = await api.get(`/categories/getallcategories`,{headers:{
+            'Authorization': `Bearer ${localStorage.getItem("accessToken")}`
+        }});
+
         console.log('Categories fetched:', response.data);
         return response.data.data;
     } catch (error) {
@@ -29,7 +50,9 @@ export const getAllCategories = async () => {
 
 export const getCategoryById = async (categoryId) => {
     try {
-        const response = await api.get(`/categories/getcategory/${categoryId}`);
+        const response = await api.get(`/categories/getcategory/${categoryId}`,{headers:{
+            'Authorization': `Bearer ${localStorage.getItem("accessToken")}`
+        }});
         console.log('Category fetched:', response.data);
         return response.data.data;
     } catch (error) {
@@ -41,7 +64,9 @@ export const getCategoryById = async (categoryId) => {
 
 export const deleteCategory = async (categoryId) => {
     try {
-        const response = await api.delete(`/categories/deletecategory/${categoryId}`);
+        const response = await api.delete(`/categories/admin/deletecategory/${categoryId}`,{headers:{
+            'Authorization': `Bearer ${localStorage.getItem("accessToken")}`
+        }});
         toast.success('Category deleted successfully!');
         console.log('Category deleted:', response);
         return response.data;
